+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Ptr-of operator (`&&`) can now make a pointer to a named function. (Instead, `__nativeFun funName` is no longer supported.)
 - Add `Std.Ptr.FunPtr.invoke`
 - Add opaque types
+- Support generic records
 
 ### CLI Tool